module9-task1: big refactoring (part 4)
