+++ Readme.md
@@ -1,7 +1,7 @@
 # Личный проект «Что посмотреть»
 
 * Студент: [Владислав Аткишкин](https://up.htmlacademy.ru/univer-nodejs-urfu/1/user/840933).
-* Наставник: `Глеб Клецков`.
+* Наставник: [Глеб Клецков](https://htmlacademy.ru/profile/id1487865).
 
 ---
