+++ package.json
@@ -17,8 +17,7 @@
     "clean": "rimraf dist",
     "ts": "ts-node",
     "doc": "npx redoc-cli serve specification/specification.yaml",
-    "mockServer": "json-server ./mocks/mock-server-data.json --port 8088",
-    "cli": "npm run ts src/main.ts --"
+    "mockServer": "json-server ./mocks/mock-server-data.json --port 8888"
   },
   "devDependencies": {
     "@types/convict": "6.1.1",
