+++ queries.http
@@ -141,6 +141,29 @@ Authorization: Bearer eyJhbGciOiJIUzI1NiJ9.eyJlbWFpbCI6InZsYWRpa0B2ay5jb20iLCJpZ
 ###
 
 # Пользователи
+## Создание нового пользователя (без аватарки => будет дефолтной)
+POST http://localhost:8000/users/register HTTP/1.1
+Content-Type: multipart/form-data; boundary=DataToLogin
+
+--DataToLogin
+Content-Disposition: form-data; name="email"
+Content-Type: text/plain
+
+evik@vk.com
+--DataToLogin
+Content-Disposition: form-data; name="name"
+Content-Type: text/plain
+
+evik
+--DataToLogin
+Content-Disposition: form-data; name="password"
+Content-Type: text/plain
+
+qwerty123
+--DataToLogin--
+
+###
+
 ## Создание нового пользователя
 POST http://localhost:8000/users/register HTTP/1.1
 Content-Type: multipart/form-data; boundary=DataToLogin
