+++ src/constants/commentCount.constants.ts
@@ -1,0 +1,1 @@
+export const MAX_COMMENTS_COUNT = 50;
