+++ src/constants/crypto.constants.ts
@@ -1,0 +1,1 @@
+export const JWT_ALGORITHM = 'HS256';
