+++ src/constants/moviesCount.constant.ts
@@ -1,0 +1,1 @@
+export const MAX_MOVIES_COUNT = 60;
