+++ src/constants/staticImages.constant.ts
@@ -1,0 +1,3 @@
+import fs from 'fs/promises';
+
+export const STATIC_IMAGES = await fs.readdir('./static');
