+++ src/modules/comment/comment.route.ts
@@ -1,5 +1,3 @@
-export const MAX_COMMENTS_COUNT = 50;
-
 export enum CommentRoute {
   Root = '/',
 }
