+++ src/modules/movie/movie.route.ts
@@ -1,7 +1,3 @@
-import fs from 'fs/promises';
-
-export const MAX_MOVIES_COUNT = 60;
-
 export enum MovieRoute {
   Root = '/',
   Create = '/create',
@@ -9,5 +5,3 @@ export enum MovieRoute {
   Promo = '/promo',
   Comments = '/:movieId/comments'
 }
-
-export const STATIC_IMAGES = await fs.readdir('./static');
