+++ src/modules/user/user.models.ts
@@ -6,6 +6,4 @@ export enum UserRoute {
   Static = '/static'
 }
 
-export const JWT_ALGORITHM = 'HS256';
-
 export const DEFAULT_AVATAR_FILE_NAME = 'default-avatar.png';
