+++ src/utils/db.ts
@@ -1,7 +1,7 @@
-export function getDBConnectionURI(username: string,
+export const getDBConnectionURI = (
+  username: string,
   password: string,
   host: string,
   port: number,
-  databaseName: string): string {
-  return `mongodb://${username}:${password}@${host}:${port}/${databaseName}?authSource=admin`;
-}
+  databaseName: string
+): string => `mongodb://${username}:${password}@${host}:${port}/${databaseName}?authSource=admin`;
